gerenciamento-de-tarefas-mongoBD/src/requirements.txt
pandas==1.4.4
numpy==1.23.2
pymongo==4.2.0
python-dateutil==2.8.2